+++ banking/src/main/java/com/project88/banking/repository/IBillRepository.java
@@ -12,7 +12,7 @@ import com.project88.banking.entity.User;
 public interface IBillRepository extends JpaRepository<Bill, Integer> {
 
 	@Query("Select b FROM Bill b WHERE b.user.userID = :userId")
-	List<Bill> findBillByUserId(@Param("userId") short userId);
+	List<Bill> findBillByUserId(@Param("userId") Long userId);
     
     @Query("SELECT b.user FROM Bill b WHERE b.billId = :billId")
     User findUserByBillId(@Param("billId") int billId);
