+++ banking/src/main/java/com/project88/banking/service/IBillService.java
@@ -6,7 +6,7 @@ import com.project88.banking.entity.Bill;
 
 public interface IBillService {
 
-	List<Bill> findBillByUserId(short userId);
+	List<Bill> findBillByUserId(Long userId);
 
 	void payBill(int billId);
 }
