+++ banking/src/main/java/com/project88/banking/service/IDepositService.java
@@ -4,6 +4,6 @@ import com.project88.banking.dto.DepositDTO;
 
 public interface IDepositService {
 
-	void createDeposit(DepositDTO form, short userId);
+	void createDeposit(DepositDTO form, Long userId);
 
 }
